requirements.txt
aiohttp
zeroconf
pycryptodome
protobuf